Implement Exploratory Daemon (#19)
Implement Exploratory Daemon.

See cmudb/exploration/README.md for more details.

This commit adds the functionality to take online copies of a
postgres database. The goal of the copy is to be able to run fast
but not necessarily have the most up to date data. The process of
taking a copy is the following:

1. Run checkpoint on original database (This is not strictly
required).
2. Take a ZFS snapshot and clone of pgdata directory.
3. Run pg_resetwal on ZFS copy. This will prune the log to some
consistent point which makes starting the new instance much faster.
However it also results in some lost data.
4. Start exploratory docker container which uses the copied pgdata
directory to start a postgres instance.

The process of tearing down an exploratory instance is the
following:

1. Stop postgres instance.
2. Shutdown docker container.
3. Delete copied data.

The daemon needs to be run on a system with an existing postgres
instance and ZFS installed.

The current implementation just takes a single copy, runs some
random queries, and tears down the copy. In the future this
functionality will be daemonized to allow on demand copies to
be made.